FlickSuggestViewのリファクタリング
* 複数の型を別ファイルに分けた

* FlickedKeyModel側がViewを作成していた問題を修正

* ThemeDataに新たにsuggestLabelTextColorを追加し、文字色を変えやすく

* コメントを追加
