+++ AzooKeyCore/Sources/KeyboardThemes/ThemeData.swift
@@ -24,9 +24,10 @@ public struct ThemeData<ApplicationExtension: ApplicationSpecificTheme>: Codable
     public var specialKeyFillColor: ColorData
     public var pushedKeyFillColor: ColorData   // 自動で設定する
     public var suggestKeyFillColor: ColorData?  // 自動で設定する
-    public var keyShadow: ThemeShadowData<ColorData>?
+    public var suggestLabelTextColor: ColorData?        // 設定は露出させない
+    public var keyShadow: ThemeShadowData<ColorData>?   // 設定は露出させない
 
-    public init(id: Int? = nil, backgroundColor: ColorData, picture: ThemePicture, textColor: ColorData, textFont: ThemeFontWeight, resultTextColor: ColorData, resultBackgroundColor: ColorData, borderColor: ColorData, borderWidth: Double, normalKeyFillColor: ColorData, specialKeyFillColor: ColorData, pushedKeyFillColor: ColorData, suggestKeyFillColor: ColorData? = nil, keyShadow: ThemeShadowData<ColorData>? = nil) {
+    public init(id: Int? = nil, backgroundColor: ColorData, picture: ThemePicture, textColor: ColorData, textFont: ThemeFontWeight, resultTextColor: ColorData, resultBackgroundColor: ColorData, borderColor: ColorData, borderWidth: Double, normalKeyFillColor: ColorData, specialKeyFillColor: ColorData, pushedKeyFillColor: ColorData, suggestKeyFillColor: ColorData? = nil, suggestLabelTextColor: ColorData? = nil, keyShadow: ThemeShadowData<ColorData>? = nil) {
         self.id = id
         self.backgroundColor = backgroundColor
         self.picture = picture
@@ -40,6 +41,7 @@ public struct ThemeData<ApplicationExtension: ApplicationSpecificTheme>: Codable
         self.specialKeyFillColor = specialKeyFillColor
         self.pushedKeyFillColor = pushedKeyFillColor
         self.suggestKeyFillColor = suggestKeyFillColor
+        self.suggestLabelTextColor = suggestLabelTextColor
         self.keyShadow = keyShadow
     }
 
@@ -57,6 +59,7 @@ public struct ThemeData<ApplicationExtension: ApplicationSpecificTheme>: Codable
         case specialKeyFillColor
         case pushedKeyFillColor
         case suggestKeyFillColor
+        case suggestLabelTextColor
         case keyShadow
     }
 
@@ -76,6 +79,7 @@ public struct ThemeData<ApplicationExtension: ApplicationSpecificTheme>: Codable
         self.specialKeyFillColor = try container.decode(ColorData.self, forKey: .specialKeyFillColor)
         self.pushedKeyFillColor = try container.decode(ColorData.self, forKey: .pushedKeyFillColor)
         self.suggestKeyFillColor = try? container.decode(ColorData?.self, forKey: .suggestKeyFillColor)
+        self.suggestLabelTextColor = try? container.decode(ColorData?.self, forKey: .suggestLabelTextColor)
         self.keyShadow = try? container.decode(ThemeShadowData<ColorData>?.self, forKey: .keyShadow)
     }
