+++ AzooKeyCore/Sources/KeyboardViews/View/FlickKeyboard/KeyView/FlickKeyModel.swift
@@ -11,7 +11,6 @@ import Foundation
 import KeyboardThemes
 import SwiftUI
 
-// M：基本は変わらない
 struct FlickKeyModel<Extension: ApplicationSpecificKeyboardViewExtension>: FlickKeyModelProtocol {
     static var delete: Self { Self(labelType: .image("delete.left"), pressActions: [.delete(1)], longPressActions: .init(repeat: [.delete(1)]), flickKeys: [
         .left: FlickedKeyModel(
